@@ -8,7 +8,7 @@
     "url": "git@github.com:SoftwareAG/cumulocity-dynamic-mapper.git"
   },
   "scripts": {
-    "start": "ng serve --host 0.0.0.0",
+    "start": "ng serve",
     "start:admin": "npm start -- --shell administration",
     "build": "ng build",
     "deploy": "ng deploy",
@@ -19,49 +19,48 @@
   "keywords": [],
   "license": "ISC",
   "dependencies": {
-    "@angular/animations": "^16.2.0",
-    "@angular/cdk": "^16.2.11",
-    "@angular/common": "^16.2.0",
-    "@angular/compiler": "^16.2.0",
-    "@angular/core": "^16.2.0",
-    "@angular/forms": "^16.2.0",
-    "@angular/platform-browser": "^16.2.0",
-    "@angular/platform-browser-dynamic": "^16.2.0",
-    "@angular/router": "^16.2.0",
-    "@angular/upgrade": "^16.2.0",
-    "@c8y/bootstrap": "1019.23.8",
-    "@c8y/client": "1019.23.8",
-    "@c8y/ngx-components": "1019.23.8",
-    "@c8y/style": "1019.23.8",
+    "@angular/animations": "^17.3.0",
+    "@angular/common": "^17.3.0",
+    "@angular/compiler": "^17.3.0",
+    "@angular/core": "^17.3.0",
+    "@angular/forms": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
+    "@angular/platform-server": "^17.3.0",
+    "@angular/router": "^17.3.0",
+    "@angular/upgrade": "^17.3.0",
+    "@c8y/bootstrap": "^1020.0.0",
+    "@c8y/ngx-components": "^1020.0.0",
+    "@c8y/options": "^1020.0.15",
     "@ngx-translate/core": "14.0.0",
     "chart.js": "^4.4.1",
     "css-loader": "^6.10.0",
     "jsonata": "^2.0.3",
-    "ngx-bootstrap": "11.0.2",
+    "ngx-bootstrap": "^12.0.0",
     "postcss-loader": "^8.1.0",
     "rxjs": "~7.8.1",
     "style-loader": "^3.3.4",
     "vanilla-jsoneditor": "^0.23.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.0",
+    "@angular-devkit/build-angular": "^17.3.8",
     "@angular-eslint/eslint-plugin": "^17.1.1",
     "@angular-eslint/eslint-plugin-template": "^17.1.1",
     "@angular-eslint/template-parser": "^17.1.1",
-    "@angular/compiler-cli": "^16.2.0",
-    "@angular/language-service": "^16.2.0",
-    "@angular/localize": "^16.2.0",
-    "@angular/service-worker": "^16.2.0",
-    "@c8y/devkit": "1019.23.8",
-    "@c8y/widget-plugin": "1019.23.8",
+    "@angular/compiler-cli": "^17.3.11",
+    "@angular/language-service": "^17.3.0",
+    "@angular/localize": "^17.3.0",
+    "@angular/service-worker": "^17.3.0",
+    "@c8y/devkit": "^1020.0.15",
+    "@c8y/package-blueprint": "1020.0.15",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
-    "cumulocity-cypress": "^0.2.4",
+    "cumulocity-cypress": "^0.4.6",
     "cypress": "^13.6.4",
     "cypress-slow-down": "^1.3.1",
     "eslint-plugin-import": "^2.29.1",
     "file-loader": "^6.2.0",
-    "prettier": "^3.1.1",
-    "typescript": "~5.1.3"
+    "prettier": "^3.2.5",
+    "typescript": "~5.4.2"
   }
 }
