@@ -31,5 +31,5 @@ public enum Operation {
     RESET_STATUS_MAPPING,
     REFRESH_NOTIFICATIONS_SUBSCRIPTIONS,
     DEBUG_MAPPING,
-
+    SNOOP_MAPPING,
 }
