@@ -98,11 +98,13 @@ export enum Operation {
   RELOAD_MAPPINGS = 'RELOAD_MAPPINGS',
   RESET_STATUS_MAPPING = 'RESET_STATUS_MAPPING',
   REFRESH_NOTIFICATIONS_SUBSCRIPTIONS = 'REFRESH_NOTIFICATIONS_SUBSCRIPTIONS',
-  DEBUG_MAPPING = 'DEBUG_MAPPING'
+  DEBUG_MAPPING = 'DEBUG_MAPPING',
+  SNOOP_MAPPING = 'SNOOP_MAPPING'
 }
 
 export enum StatusEventTypes {
   STATUS_CONNECTOR_EVENT_TYPE = 'd11r_connectorStatusEvent',
+  STATUS_MAPPING_CHANGED_EVENT_TYPE = 'd11r_mappingChangedEvent',
   STATUS_SUBSCRIPTION_EVENT_TYPE = 'd11r_subscriptionEvent',
   STATUS_NOTIFICATION_EVENT_TYPE = 'd11r_notificationStatusEvent',
   ALL = 'ALL'
