@@ -1,5 +1,5 @@
 import './i18n';
-import { applyOptions, loadOptions, loginOptions } from '@c8y/bootstrap';
+import { applyOptions, loadOptions } from '@c8y/bootstrap';
 
 const barHolder: HTMLElement = document.querySelector('body > .init-load');
 export const removeProgress = () => barHolder && barHolder.parentNode.removeChild(barHolder);
@@ -8,8 +8,7 @@ applicationSetup();
 
 async function applicationSetup() {
   const options = await applyOptions({
-    ...(await loadOptions()),
-    ...((await loginOptions()) as object)
+    ...(await loadOptions())
   });
 
   const mod = await import('./bootstrap');
