@@ -35,25 +35,15 @@ export class MappingTabFactory implements TabFactory {
 
     const tabs: Tab[] = [];
     if (this.router.url.match(/sag-ps-pkg-dynamic-mapping/g)) {
-      if (feature?.userHasMappingAdminRole) {
-        tabs.push({
-          path: 'sag-ps-pkg-dynamic-mapping/configuration',
-          priority: 930,
-          label: 'Connector',
-          icon: 'cog',
-          orientation: 'horizontal'
-        } as Tab);
-      }
-
-      // TODO investigate why hide does not work, potential bug in component
-      //   tabs.push({
-      //     path: 'sag-ps-pkg-dynamic-mapping/configuration',
-      //     priority: 930,
-      //     label: 'Connector',
-      //     icon: 'cog',
-      //     orientation: 'horizontal',
-      //     hide: !feature?.userHasMappingAdminRole
-      //   } as Tab);
+      // if (feature?.userHasMappingAdminRole) {
+      tabs.push({
+        path: 'sag-ps-pkg-dynamic-mapping/configuration',
+        priority: 930,
+        label: 'Connector',
+        icon: 'cog',
+        orientation: 'horizontal'
+      } as Tab);
+      // }
 
       tabs.push({
         path: 'sag-ps-pkg-dynamic-mapping/mappings/inbound',
@@ -73,16 +63,6 @@ export class MappingTabFactory implements TabFactory {
         } as Tab);
       }
 
-      // TODO investigate why hide does not work, potential bug in component
-      //   tabs.push({
-      //     path: 'sag-ps-pkg-dynamic-mapping/mappings/outbound',
-      //     priority: 910,
-      //     label: 'Mapping outbound',
-      //     icon: 'swipe-left',
-      //     orientation: 'horizontal',
-      //     hide: !feature?.outputMappingEnabled
-      //   } as Tab);
-
       tabs.push({
         path: 'sag-ps-pkg-dynamic-mapping/testing',
         priority: 900,
@@ -99,25 +79,15 @@ export class MappingTabFactory implements TabFactory {
         orientation: 'horizontal'
       } as Tab);
 
-      if (feature?.userHasMappingAdminRole) {
-        tabs.push({
-          path: 'sag-ps-pkg-dynamic-mapping/extension',
-          priority: 880,
-          label: 'Processor extension',
-          icon: 'plugin',
-          orientation: 'horizontal'
-        } as Tab);
-      }
-
-      // TODO investigate why hide does not work, potential bug in component
-      //   tabs.push({
-      //     path: 'sag-ps-pkg-dynamic-mapping/extension',
-      //     priority: 880,
-      //     label: 'Processor extension',
-      //     icon: 'plugin',
-      //     orientation: 'horizontal'
-      //     hide: !feature?.userHasMappingAdminRole
-      //   } as Tab);
+      // if (feature?.userHasMappingAdminRole) {
+      tabs.push({
+        path: 'sag-ps-pkg-dynamic-mapping/extension',
+        priority: 880,
+        label: 'Processor extension',
+        icon: 'plugin',
+        orientation: 'horizontal'
+      } as Tab);
+      //  }
     }
     return tabs;
   }
