@@ -29,6 +29,7 @@ export interface MappingSubstitution {
 }
 
 export interface Mapping {
+  [x: string]: any;
   name: string;
   id: string;
   ident: string;
@@ -66,7 +67,7 @@ export interface MappingEnriched {
   deployedToConnectors?: ConnectorConfiguration[];
 }
 
-export interface MappingSubscribed {
+export interface MappingDeployed {
   ident: string;
   deployedToConnectors?: ConnectorConfiguration[];
 }
